+++ src/main/java/net/xuyifei/lolipickaxe/common/entity/ModEntities.java
@@ -12,6 +12,7 @@ import net.neoforged.neoforge.registries.DeferredRegister;
 import net.xuyifei.lolipickaxe.LoliPickaxe;
 import net.xuyifei.lolipickaxe.common.entity.render.EntityLoliRenderer;
 import net.xuyifei.lolipickaxe.common.entity.render.ModelLoli;
+import net.xuyifei.lolipickaxe.common.entity.render.PrimedLoliBuffAttackTNTEntityRenderer;
 
 import java.util.function.Supplier;
 
@@ -21,6 +22,8 @@ public class ModEntities {
     public static final Supplier<EntityType<EntityLoli>> LOLI_ENTITY = ENTITIES.register("loli_entity", () -> EntityType.Builder.of(EntityLoli::new, MobCategory.MONSTER)
             .sized(0.6F, 1.8F)
             .build("loli_entity"));
+
+    public static final Supplier<EntityType<PrimedLoliBuffAttackTNTEntity>> LOLI_BUFF_ATTACK_TNT = ENTITIES.register("loli_buff_attack_tnt", () -> EntityType.Builder.<PrimedLoliBuffAttackTNTEntity>of(PrimedLoliBuffAttackTNTEntity::new, MobCategory.MISC).fireImmune().sized(0.98F, 0.98F).eyeHeight(0.15F).clientTrackingRange(10).updateInterval(10).build("loli_buff_attack_tnt"));
 
     public static void register(IEventBus bus) {
         ENTITIES.register(bus);
@@ -41,6 +44,7 @@ public class ModEntities {
     public static void onClientSetup(FMLClientSetupEvent event) {
         event.enqueueWork(() -> {
             EntityRenderers.register(ModEntities.LOLI_ENTITY.get(), EntityLoliRenderer::new);
+            EntityRenderers.register(ModEntities.LOLI_BUFF_ATTACK_TNT.get(), PrimedLoliBuffAttackTNTEntityRenderer::new);
         });
     }
 }